Few Suggestions
1) Moved odometry to a separate class and made it a public member variable of DriveTrain so its methods can be directly accessed through DriveTrain and DriveTrain.java can remain very simple.
2) Removed a few helper functions that aren't needed.
3) Other changes self explanatory/described in messenger
